chore(apr): remove dead GPU APR inference shortcut functions (#170)
Per bug #170, the legacy `run_apr_inference_gpu_q4k` and
`run_apr_inference_gpu` functions produced garbage output and were
already disabled — callers in cli/mod.rs fall through to
`dispatch_inference → run_apr_inference → from_apr →
OwnedQuantizedModelCuda`, which is the same path GGUF uses and is
known-correct.

Scope:
- Deleted crates/aprender-serve/src/cli/apr_gpu_inference.rs (612 lines).
- Removed its `include!()` from cli/inference.rs.
- Removed dead `print_gpu_debug_weights` (~105 lines) from
  cli/apr_inference.rs.
- Preserved `argmax` and `sample_with_temperature` helpers by moving
  them into cli/apr_inference.rs — they are still live, used by
  api/apr_q4k_scheduler.rs which imports
  `crate::cli::inference::{argmax, sample_with_temperature}`.
- Updated stale #170 comment in cli/mod.rs to reflect the removal.

Net: -715 / +38 lines across 4 files. CUDA path behavior unchanged.
No functional regressions — the removed code was already unreachable.
