+++ crates/aprender-serve/src/cli/inference.rs
@@ -410,4 +410,3 @@ pub fn run_gguf_inference(
 
 include!("gguf_inference_gpu.rs");
 include!("apr_inference.rs");
-include!("apr_gpu_inference.rs");
